arrow single test replaced with multiple test cases specific to just each arrow's symbol test cases for all arrows together with "use" elements for whats not in the symbol which can be viewed (e.g. in a web browser) test case viewable in a web browser that includes descriptions for what the arrows mean
arrow single test replaced with multiple
test cases specific to just each arrow's symbol
test cases for all arrows together with "use" elements for whats not in the symbol which can be viewed (e.g. in a web browser)
test case viewable in a web browser that includes descriptions for what the arrows mean

arrow single test replaced with multiple
test cases specific to just each arrow's symbol
test cases for all arrows together with "use" elements for whats not in the symbol which can be viewed (e.g. in a web browser)
test case viewable in a web browser that includes descriptions for what the arrows mean